requirements.txt
matplotlib==3.10.1
numpy==2.2.4
Pillow==11.1.0
reedsolo==1.7.0
skimage==0.0
torch==2.2.2+cu118
torchvision==0.17.2
tqdm==4.66.1
